Add support for iOS 13 scene lifecycle
This commit adds support for the iOS 13 scene lifecycle by implementing the `application(_:configurationForConnecting:options:)` and `application(_:didDiscardSceneSessions:)` methods in the `AppDelegate` class. It also adds the `@available(iOS 13.0, *)` attribute to these methods. Additionally, a new private method called `checkFontFamily()` is added to the `AppDelegate` class for checking font names after adding them. In the `SceneDelegate` class, the `@available(iOS 13.0, *)` attribute is added to indicate compatibility with iOS 13 or later versions.