Add LoginViewModel for handling login functionality
This commit adds a new file called LoginViewModel.swift which contains the implementation of the LoginViewModel class. The LoginViewModel class is responsible for handling the login functionality in the app.

The LoginViewModel class conforms to the ViewModel protocol and implements the requestLogin method from the RequestService protocol. It also defines several closure properties for handling different scenarios such as successful login, failed login, no refresh token found, and no access token found.

Inside the requestLogin method, it makes a network request using NetworkManager.shared.request to sign in with the provided email and password. Upon success, it saves the access token and refresh token to UserDefault using UserDefault().saveAccessToken(data:) and UserDefault().saveRefreshToken(data:) respectively. It then starts a timer using TokenRefresher.shared.startTimer().

If there is an error during the network request, it checks if it's a server error or not. If it's a server error, it extracts the error details from the response and calls self.loginFailed? with that detail. Otherwise, it calls self.loginFailed? with the localized description of the error.

These changes enable users to log in to their accounts successfully and handle any errors that may occur during login.