Add setupUI method to ForgotPasswordViewController
This commit adds a new method called `setupUI` to the `ForgotPasswordViewController` class. The `setupUI` method is responsible for setting up the user interface of the view controller. In this case, it initializes and configures a notification view with a blur effect and adds it to the navigation controller's view.

The `setupNotificationView` private method is also added, which creates and configures the individual components of the notification view, such as the title, message, and image.

These changes improve the overall user experience by providing visual feedback when resetting a password.