Add Keychain.swift for saving, loading, and deleting data from keychain
This commit adds a new file called Keychain.swift which contains the implementation of saving, loading, and deleting data from the keychain. The `save` function saves data to the keychain using a given key and data. The `load` function retrieves data from the keychain using a given key. The `delete` function deletes data from the keychain using a given key.

These changes allow for secure storage of sensitive information in the app.