Add ObjectSavable.swift to support file manager functionality.
This commit adds the ObjectSavable.swift file, which provides support for saving and retrieving objects using the FileManager. This file includes methods for encoding and decoding objects to and from data, as well as saving and loading objects to and from disk.

The ObjectSavable protocol defines the required methods for an object to be saved and loaded using the FileManager. Objects that conform to this protocol can easily be saved and retrieved using the FileManager's writeData(_:to:options:) and readData(from:options:) methods.

This addition enhances the functionality of the NimbleSurvey app by providing a convenient way to save and load objects using the FileManager.