Add refreshTokenFailure callback to LoginViewModel
This commit adds a new public variable `refreshTokenFailure` to the `LoginViewModel` class. It also modifies the existing code in the `requestAccessToken` and `requestRefreshToken` methods to handle cases where no refresh token or access token is found. If no refresh token is found, the `noRefreshTokenFound` callback is triggered. If no access token is found, the `noAccessTokenFound` callback is triggered.